fix compiling waring
git-svn-id: http://192.168.20.11/sequoiadb/trunk@33040 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/client/clientcpp.cpp
@@ -5673,7 +5673,7 @@ error :
 
       reply = ( const MsgOpReply * )( _pReceiveBuffer ) ;
       if ( reply->numReturned > 0 &&
-           reply->header.messageLength > ossRoundUpToMultipleX( sizeof(MsgOpReply), 4 ) )
+           (UINT32)reply->header.messageLength > ossRoundUpToMultipleX( sizeof(MsgOpReply), 4 ) )
       {
          // get reply bson from received msg
          const CHAR* bsonBuf = _pReceiveBuffer + sizeof( MsgOpReply ) ;
